refactor law display into progression module
