+++ src/features/progression/index.js
@@ -7,3 +7,10 @@ export {
 } from './ui/realm.js';
 
 export { advanceRealm, checkLawUnlocks, awardLawPoints } from './mutators.js';
+
+export {
+  updateLawsDisplay,
+  renderLawSelection,
+  renderSkillTrees,
+  mountLawDisplay,
+} from './ui/lawDisplay.js';
